+++ projects/test-servers/nuxt-server/server/api/[...remult].ts
@@ -1,11 +1,11 @@
 import { remultApi } from 'remult/remult-nuxt'
 import { Task } from '~/shared/Task.js'
-import { someRoutes } from '../../../shared/modules/someRoutes/serverDist.js'
+import { someRoutes } from '../../../shared/modules/someRoutes/server.js'
 
 export const api = remultApi({
   entities: [Task],
   admin: true,
-  modules: [someRoutes],
+  modules: [someRoutes as any],
 })
 
 export default defineEventHandler(api)
